requirements.txt
PyPDF2==3.0.1
langchain==0.3.27
faiss-cpu
sentence-transformers==5.1.0
langchain-community==0.3.29
accelerate==1.10.1
autoawq==0.2.9
#intel_extension_for_pytorch==2.5.0
optimum==1.27.0
auto-gptq==0.7.1
huggingface_hub
docling==2.52.0